Layer Supertype Pattern #1302
Implemented the Layer Supertype Pattern in Service Layer, updated README file and PUML class diagram to reflect the new changes.
