+++ service-layer/src/main/java/com/iluwatar/servicelayer/spell/SpellDao.java
@@ -31,6 +31,5 @@ import com.iluwatar.servicelayer.common.Dao;
  */
 public interface SpellDao extends Dao<Spell> {
 
-  Spell findByName(String name);
 
 }
