+++ service-layer/src/main/java/com/iluwatar/servicelayer/spell/SpellDaoImpl.java
@@ -37,25 +37,4 @@ import org.hibernate.query.Query;
  */
 public class SpellDaoImpl extends DaoBaseImpl<Spell> implements SpellDao {
 
-  @Override
-  public Spell findByName(String name) {
-    Transaction tx = null;
-    Spell result;
-    try (var session = getSessionFactory().openSession()) {
-      tx = session.beginTransaction();
-      CriteriaBuilder criteriaBuilder = session.getCriteriaBuilder();
-      CriteriaQuery<Spell> builderQuery = criteriaBuilder.createQuery(Spell.class);
-      Root<Spell> root = builderQuery.from(Spell.class);
-      builderQuery.select(root).where(criteriaBuilder.equal(root.get("name"), name));
-      Query<Spell> query = session.createQuery(builderQuery);
-      result = query.uniqueResult();
-      tx.commit();
-    } catch (Exception e) {
-      if (tx != null) {
-        tx.rollback();
-      }
-      throw e;
-    }
-    return result;
-  }
 }
