+++ service-layer/src/main/java/com/iluwatar/servicelayer/spellbook/SpellbookDao.java
@@ -31,6 +31,4 @@ import com.iluwatar.servicelayer.common.Dao;
  */
 public interface SpellbookDao extends Dao<Spellbook> {
 
-  Spellbook findByName(String name);
-
 }
