+++ service-layer/src/main/java/com/iluwatar/servicelayer/spellbook/SpellbookDaoImpl.java
@@ -37,26 +37,4 @@ import org.hibernate.query.Query;
  */
 public class SpellbookDaoImpl extends DaoBaseImpl<Spellbook> implements SpellbookDao {
 
-  @Override
-  public Spellbook findByName(String name) {
-    Transaction tx = null;
-    Spellbook result;
-    try (var session = getSessionFactory().openSession()) {
-      tx = session.beginTransaction();
-      CriteriaBuilder criteriaBuilder = session.getCriteriaBuilder();
-      CriteriaQuery<Spellbook> builderQuery = criteriaBuilder.createQuery(Spellbook.class);
-      Root<Spellbook> root = builderQuery.from(Spellbook.class);
-      builderQuery.select(root).where(criteriaBuilder.equal(root.get("name"), name));
-      Query<Spellbook> query = session.createQuery(builderQuery);
-      result = query.uniqueResult();
-      tx.commit();
-    } catch (Exception e) {
-      if (tx != null) {
-        tx.rollback();
-      }
-      throw e;
-    }
-    return result;
-  }
-
 }
