+++ service-layer/src/main/java/com/iluwatar/servicelayer/wizard/WizardDao.java
@@ -31,6 +31,4 @@ import com.iluwatar.servicelayer.common.Dao;
  */
 public interface WizardDao extends Dao<Wizard> {
 
-  Wizard findByName(String name);
-
 }
